+++ packages/tasks-adapter-bullmq/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @goatlab/tasks-adapter-bullmq
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/node-utils@0.11.0
+
 ## 0.2.1
 
 ### Patch Changes