+++ packages/tasks-adapter-hatchet/CHANGELOG.md
@@ -1,5 +1,12 @@
 # 0.5.20
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/node-utils@0.11.0
+
 ## 0.4.0
 
 ### Minor Changes